Move WidgetAnswerDialogFragment to its own file
